requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-limiter>=3.5.0
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
networkx>=3.4.2
scikit-learn>=1.6.1
pytest>=8.3.5
nltk>=3.9.1
torch>=2.7.0
requests>=2.32.3
pandas>=2.2.3
numpy>=2.2.5
sentence-transformers
python-louvain
pyyaml
redis>=5.0.1
torch-geometric>=2.4.0
alembic>=1.13.1
python-dotenv>=1.0.1
